pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "jaydebeapi>=1.2.3",
    "mysql-connector-python>=9.4.0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
]
